LPS-79221 Move configuration related components to corresponding bundle
