+++ modules/apps/collaboration/blogs/blogs-api/src/main/resources/com/liferay/blogs/configuration/definition/packageinfo
@@ -1,1 +1,1 @@
-version 1.0.1
+version 2.0.0
