+++ modules/apps/collaboration/blogs/blogs-web/src/main/java/com/liferay/blogs/web/internal/configuration/definition/BlogsGroupServiceConfigurationBeanDeclaration.java
@@ -12,11 +12,10 @@
  * details.
  */
 
-package com.liferay.blogs.configuration.definition;
+package com.liferay.blogs.web.internal.configuration.definition;
 
 import com.liferay.blogs.configuration.BlogsGroupServiceConfiguration;
-import com.liferay.blogs.constants.BlogsConstants;
-import com.liferay.portal.kernel.settings.definition.ConfigurationPidMapping;
+import com.liferay.portal.kernel.settings.definition.ConfigurationBeanDeclaration;
 
 import org.osgi.service.component.annotations.Component;
 
@@ -24,17 +23,12 @@ import org.osgi.service.component.annotations.Component;
  * @author Sergio González
  */
 @Component
-public class BlogsGroupServiceConfigurationPidMapping
-	implements ConfigurationPidMapping {
+public class BlogsGroupServiceConfigurationBeanDeclaration
+	implements ConfigurationBeanDeclaration {
 
 	@Override
 	public Class<?> getConfigurationBeanClass() {
 		return BlogsGroupServiceConfiguration.class;
-	}
-
-	@Override
-	public String getConfigurationPid() {
-		return BlogsConstants.SERVICE_NAME;
 	}
 
 }
