+++ modules/apps/collaboration/bookmarks/bookmarks-api/src/main/resources/com/liferay/bookmarks/configuration/definition/packageinfo
@@ -1,1 +1,1 @@
-version 1.0.1
+version 2.0.0
