+++ modules/apps/collaboration/bookmarks/bookmarks-web/src/main/java/com/liferay/bookmarks/web/internal/configuration/definition/BookmarksGroupServiceConfigurationPidMapping.java
@@ -12,7 +12,7 @@
  * details.
  */
 
-package com.liferay.bookmarks.configuration.definition;
+package com.liferay.bookmarks.web.internal.configuration.definition;
 
 import com.liferay.bookmarks.configuration.BookmarksGroupServiceConfiguration;
 import com.liferay.bookmarks.constants.BookmarksConstants;
